mem leaks not done

+++ src/here_doc.c
@@ -26,6 +26,11 @@ int	here_doc(char *str)
 	}
 	close(fd);
 	fd = open("temp.txt", O_RDONLY);
+	if (fd == -1)
+    {
+        perror("Failed to reopen temporary file");
+        return -1;
+    }
 	unlink("temp.txt");
 	ft_close_all_fds();
 	return (fd);
